Fix FloatRange annotation

+++ constraintlayout/compose/src/main/java/androidx/constraintlayout/compose/ConstrainScope.kt
@@ -373,7 +373,7 @@ class ConstrainScope internal constructor(internal val id: Any) {
      */
     fun centerHorizontallyTo(
         other: ConstrainedLayoutReference,
-        @FloatRange(from = 0.0f.toDouble(), to = 1.0f.toDouble()) bias: Float = 0.5f
+        @FloatRange(from = 0.0, to = 1.0) bias: Float = 0.5f
     ) {
         linkTo(start = other.start, end = other.end, bias = bias)
     }
@@ -385,7 +385,7 @@ class ConstrainScope internal constructor(internal val id: Any) {
      */
     fun centerVerticallyTo(
         other: ConstrainedLayoutReference,
-        @FloatRange(from = 0.0f.toDouble(), to = 1.0f.toDouble()) bias: Float = 0.5f
+        @FloatRange(from = 0.0, to = 1.0) bias: Float = 0.5f
     ) {
         linkTo(other.top, other.bottom, bias = bias)
     }
